chore: update nx
