+++ apps/mobile-e2e/project.json
@@ -1,4 +1,6 @@
 {
+  "name": "mobile-e2e",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
   "sourceRoot": "apps/mobile-e2e/src",
   "projectType": "application",
   "targets": {
