+++ apps/mobile/.storybook/tsconfig.json
@@ -3,7 +3,6 @@
   "compilerOptions": {
     "emitDecoratorMetadata": true
   },
-
-  "exclude": ["../**/*.spec.ts"],
+  "exclude": ["../**/*.spec.ts", "jest.config.ts"],
   "include": ["../src/**/*", "*.js", "*.ts", "*.tsx"]
 }
