+++ apps/server/src/utils/validateActivityTime.ts
@@ -1,10 +1,8 @@
 import { ActivitySchema } from "@routine-support/domains";
-import { FormError, SubmitErrorData } from "@routine-support/forms";
+import { FormError, SubmitErrorData } from "@routine-support/types";
 import { parseTime } from "@routine-support/utils";
 
-export const validateActivityTime = (
-  activity: ActivitySchema
-): SubmitErrorData => {
+export const validateActivityTime = (activity: ActivitySchema): SubmitErrorData => {
   const startTime = parseTime(activity.start);
   const endTime = parseTime(activity.end);
