+++ apps/server/src/utils/validateCoach.ts
@@ -1,10 +1,8 @@
 import { Coach } from "@routine-support/domains";
-import { SubmitErrorData } from "@routine-support/forms";
+import { SubmitErrorData } from "@routine-support/types";
 import { validateCoachEmail } from "./validateCoachEmail";
 
-export const validateCoach = async (
-  coach: Coach
-): Promise<SubmitErrorData | undefined> => {
+export const validateCoach = async (coach: Coach): Promise<SubmitErrorData | undefined> => {
   try {
     await validateCoachEmail(coach.email);
