+++ apps/server/src/utils/validateCoachEmail.ts
@@ -1,5 +1,5 @@
 import { CoachModel } from "@routine-support/domains";
-import { FormError } from "@routine-support/forms";
+import { FormError } from "@routine-support/types";
 
 export const validateCoachEmail = async (email: string) => {
   const isEmailValid = (await CoachModel.find({ email }).countDocuments()) === 0;
