+++ apps/server/tsconfig.app.json
@@ -4,6 +4,6 @@
     "outDir": "../../dist/out-tsc",
     "types": ["node", "express"]
   },
-  "exclude": ["**/*.spec.ts"],
+  "exclude": ["**/*.spec.ts", "jest.config.ts"],
   "include": ["**/*.ts"]
 }
