+++ apps/web-e2e/project.json
@@ -1,4 +1,5 @@
 {
+  "name": "web-e2e",
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
   "sourceRoot": "apps/web-e2e/src",
   "projectType": "application",
