+++ apps/web/.storybook/tsconfig.json
@@ -13,7 +13,8 @@
     "../**/*.spec.ts",
     "../**/*.spec.js",
     "../**/*.spec.tsx",
-    "../**/*.spec.jsx"
+    "../**/*.spec.jsx",
+    "jest.config.ts"
   ],
   "include": ["../src/**/*", "*.js"]
 }
