+++ jest.config.ts
@@ -1,3 +1,4 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
 const { getJestProjects } = require("@nrwl/jest");
 
 module.exports = {
