+++ jest.preset.js
@@ -1,3 +1,3 @@
-const nxPreset = require("@nrwl/jest/preset");
+const nxPreset = require("@nrwl/jest/preset").default;
 
 module.exports = { ...nxPreset };
