+++ libs/api-client/jest.config.ts
@@ -1,3 +1,4 @@
+/* eslint-disable */
 module.exports = {
   displayName: "api-client",
   preset: "../../jest.preset.js",
@@ -8,7 +9,7 @@ module.exports = {
   },
   testEnvironment: "node",
   transform: {
-    "^.+\\.[t|j]sx?$": "babel-jest",
+    "^.+\\.[t|j]sx?$": ["babel-jest", { presets: ["@nrwl/react/babel"] }],
   },
   moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
   coverageDirectory: "../../coverage/libs/api-client",
