--- libs/constants/jest.config.js
@@ -1,15 +1,0 @@
-module.exports = {
-  displayName: "constants",
-  preset: "../../jest.preset.js",
-  globals: {
-    "ts-jest": {
-      tsconfig: "<rootDir>/tsconfig.spec.json",
-    },
-  },
-  testEnvironment: "node",
-  transform: {
-    "^.+\\.[t|j]sx?$": "babel-jest",
-  },
-  moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
-  coverageDirectory: "../../coverage/libs/constants",
-};
