+++ libs/domains/jest.config.ts
@@ -1,5 +1,6 @@
+/* eslint-disable */
 module.exports = {
-  displayName: "utils",
+  displayName: "domains",
   preset: "../../jest.preset.js",
   globals: {
     "ts-jest": {
@@ -8,8 +9,8 @@ module.exports = {
   },
   testEnvironment: "node",
   transform: {
-    "^.+\\.[t|j]sx?$": "babel-jest",
+    "^.+\\.[t|j]sx?$": ["babel-jest", { presets: ["@nrwl/react/babel"] }],
   },
   moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
-  coverageDirectory: "../../coverage/libs/utils",
+  coverageDirectory: "../../coverage/libs/models",
 };
