+++ libs/forms/jest.config.ts
@@ -3,7 +3,7 @@ export default {
   displayName: "forms",
   preset: "../../jest.preset.js",
   transform: {
-    "^.+\\.[tj]sx?$": "babel-jest",
+    "^.+\\.[tj]sx?$": ["babel-jest", { presets: ["@nrwl/react/babel"] }],
   },
   moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
   coverageDirectory: "../../coverage/libs/forms",
