--- libs/forms/src/types/FormError.ts
@@ -1,4 +1,0 @@
-export interface FormError {
-  name: string;
-  message: string;
-}
