--- libs/forms/src/types/SubmitErrorData.ts
@@ -1,7 +1,0 @@
-import { FormError } from "./FormError";
-
-export interface SubmitErrorData {
-  isValid: boolean;
-  errors?: FormError[];
-  error?: string;
-}
