+++ libs/forms/src/types/index.ts
@@ -1,5 +1,3 @@
 export type { FieldWrapperProps } from "./FieldWrapperProps";
-export type { FormError } from "./FormError";
 export type { FormFieldProps } from "./FormFieldProps";
-export type { SubmitErrorData } from "./SubmitErrorData";
 export type { FormFieldControl } from "./FormFieldControl";
