--- libs/types/src/lib/types.spec.ts
@@ -1,7 +1,0 @@
-import { types } from "./types";
-
-describe("types", () => {
-  it("should work", () => {
-    expect(types()).toEqual("types");
-  });
-});
