+++ libs/ui-theme/project.json
@@ -1,4 +1,6 @@
 {
+  "name": "ui-theme",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
   "sourceRoot": "libs/ui-theme/src",
   "projectType": "library",
   "targets": {
@@ -11,7 +13,7 @@
     },
     "test": {
       "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/ui-theme"],
+      "outputs": ["{workspaceRoot}/coverage/libs/ui-theme"],
       "options": {
         "jestConfig": "libs/ui-theme/jest.config.ts",
         "passWithNoTests": true
