+++ libs/utils/project.json
@@ -1,5 +1,6 @@
 {
-  "root": "libs/utils",
+  "name": "utils",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
   "sourceRoot": "libs/utils/src",
   "projectType": "library",
   "targets": {
@@ -12,9 +13,9 @@
     },
     "test": {
       "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/utils"],
+      "outputs": ["{workspaceRoot}/coverage/libs/utils"],
       "options": {
-        "jestConfig": "libs/utils/jest.config.js",
+        "jestConfig": "libs/utils/jest.config.ts",
         "passWithNoTests": true
       }
     }
