+++ workspace.json
@@ -14,5 +14,6 @@
     "utils": "libs/utils",
     "web": "apps/web",
     "web-e2e": "apps/web-e2e"
-  }
+  },
+  "$schema": "./node_modules/nx/schemas/workspace-schema.json"
 }
